WHACK-22: Remove Jetty
The embedded webserver (Jetty) was used for functionality that was deemed 'unused' as early as back in 2008.

Instead of carrying over maintenance debt, let's just remove it. The previous commit has an untested upgrade, that can be used for anyone that's interested in retaining this functionality, but I don't expect that someone is (if you are interested in retaining Jetty, please raise your voice!)